Auth0 implementado

+++ src/controllers/suscribers.controller.js
@@ -15,30 +15,39 @@ const getSuscriberById = async(req, res) => {
 const applySuscription = async(req, res) => {
     let actualDate = new Date(Date.now()).toLocaleString('es-AR');
     const {id, ISBN} = req.body
-    await database.query('INSERT INTO subscribed VALUES ($1,$2,$3,$4) returning id_user', [id, ISBN, actualDate, actualDate], function(err, result, fields) {
-        if (err) {
-            res.status(400).json({error: 'Something went wrong'});
-          }else{
-            res.json({
-                message: "Subscription Added Succesfully",
-                id, 
-                ISBN
-            })
-          }
-    }); 
+    if(id != null && ISBN != null && !isNaN(id)){
+        await database.query('INSERT INTO subscribed VALUES ($1,$2,$3,$4) returning id_user', [id, ISBN, actualDate, actualDate], function(err, result, fields) {
+            if (err) {
+                res.status(400).json({error: 'Something went wrong'});
+              }else{
+                res.json({
+                    message: "Subscription Added Succesfully",
+                    id, 
+                    ISBN
+                })
+              }
+        }); 
+    }else{
+        res.status(400).json({error: 'Invalid body'});
+    }
+    
 }
 
 const removeSuscription = async(req, res) => {
     const {id, ISBN} = req.body
-    await database.query('DELETE FROM  subscribed WHERE id_user = ($1) and subscribed."ISBN" = ($2)', [id, ISBN], function(err, result, fields) {
-        if (err) {
-            res.status(400).json({error: 'Something went wrong' +err});
-          }else{
-            res.json({
-                message: "Subscription Removed Succesfully"
-            })
-          }
-    }); 
+    if(id != null && ISBN != null && !isNaN(id)){
+        await database.query('DELETE FROM  subscribed WHERE id_user = ($1) and subscribed."ISBN" = ($2)', [id, ISBN], function(err, result, fields) {
+            if (err) {
+                res.status(400).json({error: 'Something went wrong' +err});
+            }else{
+                res.json({
+                    message: "Subscription Removed Succesfully"
+                })
+            }
+        }); 
+    }else{
+        res.status(400).json({error: 'Invalid body'});
+    }
 }
 
 module.exports = {
